address review comments and refactor e2e tests
