+++ e2e-tests/playwright/tests/self-service/ee02b-gitlab-template-execution.spec.ts
@@ -1,6 +1,6 @@
 import { test, expect } from '../../fixtures/auth-context';
-import { authenticator } from 'otplib';
 import { Page } from '@playwright/test';
+import { handleGitLabLoginOnPage } from '../../utils/auth';
 
 async function handleGitLabOAuthDialog(
   page: Page,
@@ -32,9 +32,9 @@ async function handleGitLabOAuthDialog(
 
   const context = page.context();
 
-  const popup = await Promise.race([
+  const [popup] = await Promise.all([
     context.waitForEvent('page', { timeout: 3000 }).catch(() => null),
-    logInBtn.click().then(() => null),
+    logInBtn.click(),
   ]);
 
   // Wait for the click to settle — either a popup opened or the page navigated
@@ -141,228 +141,6 @@ async function handleGitLabOAuthDialog(
   );
   await page.waitForTimeout(2000);
   return 'redirected';
-}
-
-async function clickTurnstileWidget(page: Page): Promise<boolean> {
-  const widgetBox = await page.evaluate(() => {
-    // Strategy 1: find iframe by Cloudflare challenge title or src
-    const iframes = document.querySelectorAll('iframe');
-    for (const iframe of Array.from(iframes)) {
-      const title = iframe.getAttribute('title') || '';
-      const src = iframe.getAttribute('src') || '';
-      if (
-        title.includes('Cloudflare') ||
-        title.includes('challenge') ||
-        src.includes('challenges.cloudflare') ||
-        src.includes('turnstile')
-      ) {
-        const rect = iframe.getBoundingClientRect();
-        if (rect.width > 0 && rect.height > 0) {
-          return {
-            x: rect.x,
-            y: rect.y,
-            width: rect.width,
-            height: rect.height,
-            via: 'iframe',
-          };
-        }
-      }
-    }
-
-    // Strategy 2: find container with cf-turnstile-response hidden input
-    const cfInputs = document.querySelectorAll(
-      'input[name="cf-turnstile-response"]',
-    );
-    for (const input of Array.from(cfInputs)) {
-      let el: HTMLElement | null = input.parentElement;
-      while (el) {
-        const rect = el.getBoundingClientRect();
-        if (rect.width >= 30 && rect.height >= 30) {
-          return {
-            x: rect.x,
-            y: rect.y,
-            width: rect.width,
-            height: rect.height,
-            via: 'cf-input',
-          };
-        }
-        el = el.parentElement;
-      }
-    }
-
-    // Strategy 3: div with display:grid (standalone challenge page)
-    const containers = document.querySelectorAll('div[style*="display: grid"]');
-    for (const container of Array.from(containers)) {
-      const rect = container.getBoundingClientRect();
-      if (rect.width >= 200 && rect.height >= 50) {
-        return {
-          x: rect.x,
-          y: rect.y,
-          width: rect.width,
-          height: rect.height,
-          via: 'grid-div',
-        };
-      }
-    }
-    return null;
-  });
-
-  if (!widgetBox) return false;
-
-  const clickX = widgetBox.x + 30;
-  const clickY = widgetBox.y + widgetBox.height / 2;
-  console.log(
-    `[EE GitLab Test] Turnstile (${widgetBox.via}) at (${widgetBox.x}, ${widgetBox.y}), ` +
-      `size ${widgetBox.width}x${widgetBox.height}, clicking at (${clickX}, ${clickY})`,
-  );
-  await page.mouse.click(clickX, clickY);
-  await page.waitForTimeout(3000);
-  return true;
-}
-
-async function handleSecurityVerification(page: Page): Promise<void> {
-  // Wait for the page to settle before checking for challenges
-  await page.waitForLoadState('domcontentloaded').catch(() => {});
-  await page.waitForTimeout(2000);
-
-  const hasVerifyText = await page
-    .getByText(/Performing security verification/i)
-    .isVisible({ timeout: 3000 })
-    .catch(() => false);
-
-  if (!hasVerifyText) return;
-
-  console.log('[EE GitLab Test] Security verification detected, waiting...');
-  await page.waitForTimeout(2000);
-
-  for (let attempt = 0; attempt < 10; attempt++) {
-    if (await clickTurnstileWidget(page)) {
-      console.log(
-        '[EE GitLab Test] Turnstile clicked, waiting for page to update...',
-      );
-      await page.waitForTimeout(3000);
-      return;
-    }
-
-    console.log(
-      `[EE GitLab Test] Attempt ${attempt + 1}/10 — widget not clickable yet`,
-    );
-    await page.waitForTimeout(1000);
-  }
-
-  console.log(
-    '[EE GitLab Test] Turnstile not resolved after 10 attempts, proceeding...',
-  );
-}
-
-async function fillAndSubmitGitLabLogin(page: Page): Promise<void> {
-  const loginField = page.locator('#user_login');
-  console.log('[EE GitLab Test] Filling GitLab credentials...');
-  await loginField.fill(process.env.GL_USER_ID!);
-  await page.locator('#user_password').fill(process.env.GL_USER_PASS!);
-
-  const submitBtn = page
-    .locator(
-      '[data-testid="sign-in-button"], input[type="submit"][value="Sign in"], button[type="submit"]',
-    )
-    .first();
-  await Promise.all([
-    page
-      .waitForURL(/.*/, { waitUntil: 'domcontentloaded', timeout: 15000 })
-      .catch(() => {}),
-    submitBtn.click(),
-  ]);
-  await page.waitForLoadState('networkidle').catch(() => {});
-  await page.waitForTimeout(3000);
-  console.log('[EE GitLab Test] After submit, URL:', page.url());
-}
-
-async function fillAndSubmitTOTP(page: Page): Promise<void> {
-  console.log('[EE GitLab Test] TOTP 2FA required, entering code...');
-  const totp = authenticator.generate(process.env.GL_AUTHENTICATOR_SECRET!);
-  await page.locator('#user_otp_attempt').fill(totp);
-  const totpSubmit = page
-    .locator('input[type="submit"], button[type="submit"]')
-    .first();
-  await Promise.all([
-    page
-      .waitForURL(/.*/, { waitUntil: 'domcontentloaded', timeout: 15000 })
-      .catch(() => {}),
-    totpSubmit.click(),
-  ]);
-  await page.waitForLoadState('networkidle').catch(() => {});
-  await page.waitForTimeout(3000);
-  console.log('[EE GitLab Test] After 2FA submit, URL:', page.url());
-}
-
-async function handleGitLabLoginOnPage(page: Page): Promise<void> {
-  // GitLab login flow with Cloudflare Turnstile challenges:
-  //   Turnstile → login form → submit → Turnstile → login form → submit → TOTP → Turnstile → done
-  // Each step may trigger a standalone Turnstile challenge that redirects back.
-  // Loop up to 5 rounds to handle all challenges.
-  for (let round = 0; round < 5; round++) {
-    console.log(`[EE GitLab Test] Login round ${round + 1}, URL:`, page.url());
-
-    if (new URL(page.url()).hostname !== 'gitlab.com') {
-      console.log('[EE GitLab Test] Left GitLab, login complete');
-      return;
-    }
-
-    // Handle standalone Turnstile challenge page
-    await handleSecurityVerification(page);
-
-    // Check what page we're on now
-    const hasLoginForm = await page
-      .locator('#user_login')
-      .isVisible({ timeout: 5000 })
-      .catch(() => false);
-    const hasTOTP = await page
-      .locator('#user_otp_attempt')
-      .isVisible({ timeout: 2000 })
-      .catch(() => false);
-
-    if (hasTOTP && process.env.GL_AUTHENTICATOR_SECRET) {
-      await fillAndSubmitTOTP(page);
-      continue;
-    }
-
-    if (hasLoginForm) {
-      await fillAndSubmitGitLabLogin(page);
-      continue;
-    }
-
-    // Check for error messages
-    const flash = page.locator('.flash-alert, .alert-danger').first();
-    if (await flash.isVisible({ timeout: 2000 }).catch(() => false)) {
-      console.log(
-        '[EE GitLab Test] Login error:',
-        await flash.innerText().catch(() => ''),
-      );
-    }
-
-    console.log('[EE GitLab Test] No login form or TOTP found, proceeding...');
-    break;
-  }
-
-  // Handle GitLab "Authorize" page
-  await page.waitForLoadState('networkidle').catch(() => {});
-  const authorizeBtn = page
-    .getByRole('button', { name: /authorize/i })
-    .or(page.locator('input[type="submit"][value*="Authorize"]'))
-    .first();
-  if (await authorizeBtn.isVisible({ timeout: 10000 }).catch(() => false)) {
-    console.log(
-      '[EE GitLab Test] GitLab authorize page detected, clicking Authorize...',
-    );
-    await authorizeBtn.click();
-    console.log('[EE GitLab Test] Authorize clicked, waiting for callback...');
-    await page
-      .waitForURL(url => url.hostname !== 'gitlab.com', {
-        timeout: 30000,
-      })
-      .catch(() => {});
-    console.log('[EE GitLab Test] After authorize redirect, URL:', page.url());
-  }
 }
 
 const EE_TEMPLATE_URL =
